renamed Editor things to TextEditor
